Upgrade to Variation Autoencoder
Lots of changes, restructured model so it no longer uses builder functions. Added sampling layer, added custom loss.
